Plant improvements
Plants now absorb water and nutrients and use those resources to create
energy and reproduce.  Made maximum mass determined by the organism
starting mass as a child. Mad maximum energy a multiple of the mass and
an arbitrary mutate-able value. Created the dead organism which will
take the place of animals when they die.

+++ EvolutionModel/EvolutionModel/Model/PhenoTypes/Digestion/ParasiticDigestiveSystem.cs
@@ -24,7 +24,7 @@ namespace EvolutionModel.Model.PhenoTypes.Digestion
 
         private int absorbEnergyFromOrganism(Organism organism, Type HungryFor)
         {
-            int energy = organism.EnergyTotal;
+            int energy = (int)organism.EnergyTotal;
             int absorbedEnergy = 0;
             if (this.OrganismHungryFor == HungryFor)
                 absorbedEnergy = (int)(energy * FavoredOrganismDigestRatio);
